Remove console.log statements and add comments to p3.js
I checked a few of the latest TOP solutions and noticed a number of people were having issues calculating the answer to this problem with the large number 600851475143.

I figured I'd add in some comments to describe my process to help anyone that may need it. This is FAR from perfect though, so any suggestions to improve on this is more than welcome!